JAMES-2541 Move serialization util to util module
